requirements.txt
python-dotenv==1.0.1
numpy==1.26.4
tqdm==4.67.1

transformers==4.44.2
datasets==2.21.0
peft==0.12.0
accelerate==0.34.2
sentencepiece==0.2.0
safetensors==0.4.5

torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1

scikit-learn==1.5.1
pydantic==2.8.2
langchain==0.2.14
langchain-core==0.2.33
langchain-openai==0.1.22
langchain-community==0.2.12
sentence-transformers==3.0.1
openai==1.42.0
